feat : 주변병해정보 api 구현, 위도경도 float -> double, 코드정리

+++ src/main/java/com/example/servertest/main/crop/entity/DiagnosisRecord.java
@@ -21,10 +21,10 @@ public class DiagnosisRecord {
     private long userId;
 
     @Column
-    private float userLatitude;
+    private double userLatitude;
 
     @Column
-    private float userLongitude;
+    private double userLongitude;
 
     @Column
     private LocalDateTime regDate;
